refactor: move several routing functions into the proper modules
